apps/api/requirements.txt
# Core
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5

# Data processing
PyYAML>=6.0
numpy>=1.26
scikit-learn>=1.4
jsonschema>=4.21
referencing>=0.30

# Network / providers
requests>=2.31
python-dotenv>=1.0

# Dev / test
pytest>=8.0
